pyproject.toml
[tool.poetry]
name = "hubitat_uploader"
version = "0.1.0"
description = ""
authors = ["David Manuel <david@dcmanjr.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
watchdog = "^2.1.9"
python-dotenv = "^0.21.0"
requests = "^2.28.1"
Flask = "^2.2.2"


[tool.poetry.group.dev.dependencies]
objprint = "^0.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
